+++ Changelog.md
@@ -13,6 +13,7 @@
 * Fix: Checkout v1 was not applying the custom style (#5628 #5615 #5616) @dennisreimann
 * Fix: Test email with multiple recipients was crashing (#5649 #5648) @dennisreimann
 * Fix: Test webhook for payment requests (#5680) @Kukks
+* Fix: Sometimes importing a wallet file from Electrum would fail @NicolasDorier
 
 ### Improvements